bug: sign 빌드에러 수정
